renamed main entry points and removed view suffix and jsx extension
