+++ gh-pages/views/Application.js
@@ -1,9 +1,8 @@
 import React from 'react'
 
-import Header from './HeaderView'
-import Examples from './ExamplesView'
-import License from './LicenseView'
-
+import Header from './Header'
+import Examples from './Examples'
+import License from './License'
 
 export default function ReactMentions() {
   return (
