+++ gh-pages/views/Examples.js
