+++ gh-pages/views/License.js
